feat: wire up Reviews UI and fetch requests
