+++ src/components/ProductGallery/ProductGallery.js
@@ -28,15 +28,19 @@ function ProductGallery({ images }) {
       </div>
 
       <Slider dots={false} infinite={false} slidesToShow={4} slidesToScroll={4}>
-        {images.map((image, index) => (
-          <img
-            src={image?.node?.urlThumbnail}
-            className={styles.productGalleryThumbnail}
-            onClick={() => setProductIndex(index)}
-            key={`slide-image-${index}`}
-            alt={image?.node.description}
-          />
-        ))}
+        {!images.length ? (
+          <img alt='Missing product image' src='/ProductDefault.gif' />
+        ) : (
+          images.map((image, index) => (
+            <img
+              src={image?.node?.urlThumbnail}
+              className={styles.productGalleryThumbnail}
+              onClick={() => setProductIndex(index)}
+              key={`slide-image-${index}`}
+              alt={image?.node.description}
+            />
+          ))
+        )}
       </Slider>
     </div>
   );
