--- src/components/Review/index.js
@@ -1,1 +1,0 @@
-export { default as Review } from './Review';
