+++ src/components/Reviews/index.js
@@ -1,0 +1,1 @@
+export { default as Reviews } from './Reviews';
